Mild refactoring of the parsing module
